src/requirements.txt
pydantic-settings==2.2.1
streamlit==1.32.2
chromadb==0.4.24
langchain-community==0.0.29
together==1.0.1
sqlalchemy==2.0.29
alembic==1.13.1
PyJWT==2.8.0
streamlit-router==0.1.3
streamlit-cookies-manager==0.2.0